Move vrnetlab logic into VRNode
