+++ nodes/default_node.go
@@ -829,10 +829,6 @@ func (d *DefaultNode) IsHealthy(ctx context.Context) (bool, error) {
 // Lifecycle: Stop / Start
 // ---------------------------------------------------------------------------
 
-const vrnetlabVersionLabel = "vrnetlab-version"
-
-var imageTagRE = regexp.MustCompile(`^[A-Za-z0-9][A-Za-z0-9._-]*$`)
-
 // Stop parks dataplane interfaces into a dedicated network namespace and stops the container.
 func (d *DefaultNode) Stop(ctx context.Context) error {
 	cfg := d.Config()
@@ -879,7 +875,7 @@ func (d *DefaultNode) Stop(ctx context.Context) error {
 		return fmt.Errorf("node %q failed linking parking netns: %w", cfg.ShortName, err)
 	}
 
-	preStopCleanup(ctx, d)
+	preStopCleanupNamedNetns(ctx, d)
 
 	if err := d.Runtime.StopContainer(ctx, cfg.LongName, d.StopSignal); err != nil {
 		// Docker/podman may return an error while the container is already stopped (timeout, API hiccup).
@@ -1060,14 +1056,6 @@ func setEndpointsUp(ctx context.Context, endpoints []clablinks.Endpoint) error {
 	return nil
 }
 
-func preStopCleanup(ctx context.Context, d *DefaultNode) {
-	if isVrnetlabNode(ctx, d) {
-		preStopPrepareVrnetlabQcowAlias(ctx, d)
-	}
-
-	preStopCleanupNamedNetns(ctx, d)
-}
-
 func preStopCleanupNamedNetns(ctx context.Context, d *DefaultNode) {
 	// Best-effort cleanup for containers that create named network namespaces under /run/netns.
 	// We lazily unmount active nsfs mounts and remove stale entries to avoid namespace artifacts
@@ -1101,116 +1089,3 @@ func preStopCleanupNamedNetns(ctx context.Context, d *DefaultNode) {
 	}
 }
 
-func isVrnetlabNode(ctx context.Context, d *DefaultNode) bool {
-	containers, err := d.GetContainers(ctx)
-	if err != nil {
-		log.Debugf(
-			"node %q vrnetlab detection skipped: failed to get container metadata: %v",
-			d.Config().ShortName,
-			err,
-		)
-		return false
-	}
-
-	for _, container := range containers {
-		if _, ok := container.Labels[vrnetlabVersionLabel]; ok {
-			return true
-		}
-	}
-
-	return false
-}
-
-func preStopPrepareVrnetlabQcowAlias(ctx context.Context, d *DefaultNode) {
-	aliasName, ok := vrnetlabQcowAliasName(d.Config().Image)
-	if !ok {
-		log.Debugf(
-			"node %q pre-stop vrnetlab qcow alias skipped: unable to infer tag from image %q",
-			d.Config().ShortName,
-			d.Config().Image,
-		)
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	// Some vrnetlab nodes rename the original versioned qcow image after first boot and fail on
-	// subsequent starts when they try to rediscover a versioned qcow filename. If there is exactly
-	// one non-overlay qcow file in / and our alias is absent, create a hardlink alias based on the
-	// image tag.
-	cmd := fmt.Sprintf(
-		`alias="/%s"; `+
-			`[ -e "$alias" ] && exit 0; `+
-			`src=""; `+
-			`if [ -f /sros.qcow2 ] && [ "/sros.qcow2" != "$alias" ]; then `+
-			`src="/sros.qcow2"; `+
-			`else `+
-			`set -- /*.qcow2; `+
-			`if [ "$1" != "/*.qcow2" ]; then `+
-			`for f in "$@"; do `+
-			`[ "$f" = "$alias" ] && continue; `+
-			`base="${f##*/}"; `+
-			`case "$base" in *overlay*.qcow2) continue ;; esac; `+
-			`if [ -n "$src" ]; then src=""; break; fi; `+
-			`src="$f"; `+
-			`done; `+
-			`fi; `+
-			`fi; `+
-			`[ -n "$src" ] || exit 0; `+
-			`ln "$src" "$alias"`,
-		aliasName,
-	)
-
-	execCmd := clabexec.NewExecCmdFromSlice([]string{"sh", "-lc", cmd})
-	res, err := d.RunExec(ctx, execCmd)
-	if err != nil {
-		log.Warnf(
-			"node %q pre-stop vrnetlab qcow alias preparation failed: %v",
-			d.Config().ShortName,
-			err,
-		)
-		return
-	}
-
-	if res != nil && res.ReturnCode != 0 {
-		log.Warnf(
-			"node %q pre-stop vrnetlab qcow alias prep returned code %d (stderr: %s)",
-			d.Config().ShortName,
-			res.ReturnCode,
-			res.Stderr,
-		)
-	}
-}
-
-func vrnetlabQcowAliasName(image string) (string, bool) {
-	tag, ok := imageTag(image)
-	if !ok {
-		return "", false
-	}
-
-	return "clab-" + tag + ".qcow2", true
-}
-
-func imageTag(image string) (string, bool) {
-	if at := strings.LastIndex(image, "@"); at != -1 {
-		image = image[:at]
-	}
-
-	lastSlash := strings.LastIndex(image, "/")
-	lastColon := strings.LastIndex(image, ":")
-	if lastColon == -1 || lastColon < lastSlash {
-		return "", false
-	}
-
-	tag := image[lastColon+1:]
-	if tag == "" {
-		return "", false
-	}
-
-	if !imageTagRE.MatchString(tag) {
-		return "", false
-	}
-
-	return tag, true
-}
