+++ nodes/vr_node.go
@@ -6,9 +6,12 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
+	"strings"
+	"time"
 
 	"github.com/charmbracelet/log"
 	clabconstants "github.com/srl-labs/containerlab/constants"
+	clabexec "github.com/srl-labs/containerlab/exec"
 	clablinks "github.com/srl-labs/containerlab/links"
 	clabnetconf "github.com/srl-labs/containerlab/netconf"
 	clabtypes "github.com/srl-labs/containerlab/types"
@@ -16,6 +19,7 @@ import (
 )
 
 var VMInterfaceRegexp = regexp.MustCompile(`eth[1-9]\d*$`) // skipcq: GO-C4007
+var imageTagRE = regexp.MustCompile(`^[A-Za-z0-9][A-Za-z0-9._-]*$`)
 
 type VRNode struct {
 	DefaultNode
@@ -140,4 +144,105 @@ func (n *VRNode) SaveConfig(_ context.Context) (*SaveConfigResult, error) {
 	return &SaveConfigResult{
 		ConfigPath: configPath,
 	}, nil
+}
+
+// Stop prepares vrnetlab-specific state (qcow alias) and then delegates to
+// DefaultNode.Stop which parks interfaces and stops the container.
+func (vr *VRNode) Stop(ctx context.Context) error {
+	preStopPrepareVrnetlabQcowAlias(ctx, &vr.DefaultNode)
+	return vr.DefaultNode.Stop(ctx)
+}
+
+func preStopPrepareVrnetlabQcowAlias(ctx context.Context, d *DefaultNode) {
+	aliasName, ok := vrnetlabQcowAliasName(d.Config().Image)
+	if !ok {
+		log.Debugf(
+			"node %q pre-stop vrnetlab qcow alias skipped: unable to infer tag from image %q",
+			d.Config().ShortName,
+			d.Config().Image,
+		)
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
+	defer cancel()
+
+	// Some vrnetlab nodes rename the original versioned qcow image after first boot and fail on
+	// subsequent starts when they try to rediscover a versioned qcow filename. If there is exactly
+	// one non-overlay qcow file in / and our alias is absent, create a hardlink alias based on the
+	// image tag.
+	cmd := fmt.Sprintf(
+		`alias="/%s"; `+
+			`[ -e "$alias" ] && exit 0; `+
+			`src=""; `+
+			`if [ -f /sros.qcow2 ] && [ "/sros.qcow2" != "$alias" ]; then `+
+			`src="/sros.qcow2"; `+
+			`else `+
+			`set -- /*.qcow2; `+
+			`if [ "$1" != "/*.qcow2" ]; then `+
+			`for f in "$@"; do `+
+			`[ "$f" = "$alias" ] && continue; `+
+			`base="${f##*/}"; `+
+			`case "$base" in *overlay*.qcow2) continue ;; esac; `+
+			`if [ -n "$src" ]; then src=""; break; fi; `+
+			`src="$f"; `+
+			`done; `+
+			`fi; `+
+			`fi; `+
+			`[ -n "$src" ] || exit 0; `+
+			`ln "$src" "$alias"`,
+		aliasName,
+	)
+
+	execCmd := clabexec.NewExecCmdFromSlice([]string{"sh", "-lc", cmd})
+	res, err := d.RunExec(ctx, execCmd)
+	if err != nil {
+		log.Warnf(
+			"node %q pre-stop vrnetlab qcow alias preparation failed: %v",
+			d.Config().ShortName,
+			err,
+		)
+		return
+	}
+
+	if res != nil && res.ReturnCode != 0 {
+		log.Warnf(
+			"node %q pre-stop vrnetlab qcow alias prep returned code %d (stderr: %s)",
+			d.Config().ShortName,
+			res.ReturnCode,
+			res.Stderr,
+		)
+	}
+}
+
+func vrnetlabQcowAliasName(image string) (string, bool) {
+	tag, ok := imageTag(image)
+	if !ok {
+		return "", false
+	}
+
+	return "clab-" + tag + ".qcow2", true
+}
+
+func imageTag(image string) (string, bool) {
+	if at := strings.LastIndex(image, "@"); at != -1 {
+		image = image[:at]
+	}
+
+	lastSlash := strings.LastIndex(image, "/")
+	lastColon := strings.LastIndex(image, ":")
+	if lastColon == -1 || lastColon < lastSlash {
+		return "", false
+	}
+
+	tag := image[lastColon+1:]
+	if tag == "" {
+		return "", false
+	}
+
+	if !imageTagRE.MatchString(tag) {
+		return "", false
+	}
+
+	return tag, true
 }
